+++ Assets/Plugins/UnityFx.Outline/Runtime/Scripts/OutlineEffect.cs
@@ -12,29 +12,62 @@ namespace UnityFx.Outline
 	/// Post-effect script. Should be attached to camera.
 	/// </summary>
 	/// <seealso cref="OutlineLayer"/>
+	/// <seealso cref="OutlineBehaviour"/>
 	/// <seealso cref="https://willweissman.wordpress.com/tutorials/shaders/unity-shaderlab-object-outlines/"/>
+	[RequireComponent(typeof(Camera))]
 	public sealed class OutlineEffect : MonoBehaviour
 	{
 		#region data
 
-		private const string _effectName = "Outline";
-
 #pragma warning disable 0649
 
 		[SerializeField]
-		private Shader _renderColorShader;
-		[SerializeField]
-		private Shader _postProcessShader;
+		private OutlineResources _outlineResources;
 
 #pragma warning restore 0649
 
 		private List<OutlineLayer> _layers;
 		private CommandBuffer _commandBuffer;
 		private Material _renderMaterial;
+		private bool _changed;
 
 		#endregion
 
 		#region interface
+
+		/// <summary>
+		/// Gets or sets resources used by the effect implementation.
+		/// </summary>
+		public OutlineResources OutlineResources
+		{
+			get
+			{
+				return _outlineResources;
+			}
+			set
+			{
+				if (value == null)
+				{
+					throw new ArgumentNullException("OutlineResources");
+				}
+
+				if (_outlineResources != value)
+				{
+					_outlineResources = value;
+					_changed = true;
+
+					if (_renderMaterial)
+					{
+						_renderMaterial.shader = _outlineResources.RenderShader;
+					}
+
+					foreach (var layers in _layers)
+					{
+						layers.PostProcessMaterial.shader = _outlineResources.PostProcessShader;
+					}
+				}
+			}
+		}
 
 		/// <summary>
 		/// Gets or sets a <see cref="Shader"/> that renders objects outlined with a solid while color.
@@ -43,7 +76,7 @@ namespace UnityFx.Outline
 		{
 			get
 			{
-				return _renderColorShader;
+				return _outlineResources.RenderShader;
 			}
 			set
 			{
@@ -52,9 +85,10 @@ namespace UnityFx.Outline
 					throw new ArgumentNullException("RenderColorShader");
 				}
 
-				if (_renderColorShader != value)
+				if (_outlineResources.RenderShader != value)
 				{
-					_renderColorShader = value;
+					_outlineResources.RenderShader = value;
+					_changed = true;
 
 					if (_renderMaterial)
 					{
@@ -71,7 +105,7 @@ namespace UnityFx.Outline
 		{
 			get
 			{
-				return _postProcessShader;
+				return _outlineResources.PostProcessShader;
 			}
 			set
 			{
@@ -80,13 +114,14 @@ namespace UnityFx.Outline
 					throw new ArgumentNullException("PostProcessShader");
 				}
 
-				if (_postProcessShader != value)
+				if (_outlineResources.PostProcessShader != value)
 				{
-					_postProcessShader = value;
+					_outlineResources.PostProcessShader = value;
+					_changed = true;
 
 					foreach (var layers in _layers)
 					{
-						layers.PostProcessMaterial.shader = _postProcessShader;
+						layers.PostProcessMaterial.shader = value;
 					}
 				}
 			}
@@ -120,10 +155,10 @@ namespace UnityFx.Outline
 
 			if (_renderMaterial == null)
 			{
-				_renderMaterial = new Material(_renderColorShader);
+				_renderMaterial = new Material(_outlineResources.RenderShader);
 			}
 
-			var layer = new OutlineLayer(_renderMaterial, new Material(_postProcessShader));
+			var layer = new OutlineLayer(_renderMaterial, new Material(_outlineResources.PostProcessShader));
 			_layers.Add(layer);
 			return layer;
 		}
@@ -161,11 +196,10 @@ namespace UnityFx.Outline
 			if (camera)
 			{
 				_commandBuffer = new CommandBuffer();
-				_commandBuffer.name = _effectName;
+				_commandBuffer.name = OutlineRenderer.EffectName;
+				_changed = true;
 
-				FillCommandBuffer(_commandBuffer);
-
-				camera.AddCommandBuffer(CameraEvent.BeforeImageEffects, _commandBuffer);
+				camera.AddCommandBuffer(OutlineRenderer.RenderEvent, _commandBuffer);
 			}
 		}
 
@@ -175,7 +209,7 @@ namespace UnityFx.Outline
 
 			if (camera)
 			{
-				camera.RemoveCommandBuffer(CameraEvent.BeforeImageEffects, _commandBuffer);
+				camera.RemoveCommandBuffer(OutlineRenderer.RenderEvent, _commandBuffer);
 			}
 
 			if (_commandBuffer != null)
@@ -187,20 +221,28 @@ namespace UnityFx.Outline
 
 		private void Update()
 		{
-			var needUpdate = false;
-
-			foreach (var layer in _layers)
+			if (_changed)
 			{
-				if (layer.IsChanged)
-				{
-					needUpdate = true;
-					break;
-				}
+				FillCommandBuffer();
+				_changed = false;
 			}
-
-			if (needUpdate)
+			else
 			{
-				FillCommandBuffer(_commandBuffer);
+				var needUpdate = false;
+
+				foreach (var layer in _layers)
+				{
+					if (layer.IsChanged)
+					{
+						needUpdate = true;
+						break;
+					}
+				}
+
+				if (needUpdate)
+				{
+					FillCommandBuffer();
+				}
 			}
 		}
 
@@ -208,23 +250,15 @@ namespace UnityFx.Outline
 
 		#region implementation
 
-		private void FillCommandBuffer(CommandBuffer cmdbuf)
+		private void FillCommandBuffer()
 		{
-			var rtId = Shader.PropertyToID("_MainTex");
-			var rt = new RenderTargetIdentifier(rtId);
-			var dst = new RenderTargetIdentifier(BuiltinRenderTextureType.CameraTarget);
-
-			cmdbuf.BeginSample(_effectName);
-			cmdbuf.Clear();
-			cmdbuf.GetTemporaryRT(rtId, -1, -1, 0, FilterMode.Bilinear, RenderTextureFormat.R8);
-
-			foreach (var layer in _layers)
+			using (var renderer = new OutlineRenderer(_commandBuffer, BuiltinRenderTextureType.CameraTarget))
 			{
-				layer.FillCommandBuffer(cmdbuf, rt, dst);
+				foreach (var layer in _layers)
+				{
+					layer.FillCommandBuffer(renderer);
+				}
 			}
-
-			cmdbuf.ReleaseTemporaryRT(rtId);
-			cmdbuf.EndSample(_effectName);
 		}
 
 		#endregion