+++ CHANGELOG.md
@@ -3,7 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/); this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.1.0] - Unreleased
+## [0.2.0] - 2019.08.19
+
+### Added
+- Added `OutlineBehaviour` for rendering per-object outlines.
+- Added `OutlineResources` to help initialize outline effects in runtime.
+- Added `OutlineRenderer` as low-level helper for outline rendering.
+
+## [0.1.0] - 2019.08.18
 
 ### Added
 - Initial release.